refactor: consolidate service registration into AddApplicationServices extension method
