+++ Program.cs
@@ -2,35 +2,14 @@ using Microsoft.AspNetCore.Diagnostics;
 using Microsoft.EntityFrameworkCore;
 using TodoApi.Data;
 using TodoApi.Endpoints;
-using TodoApi.Exceptions;
-using TodoApi.Services;
+using TodoApi.Extensions;
 
 var builder = WebApplication.CreateBuilder(args);
 
 builder.Logging.ClearProviders();
 builder.Logging.AddConsole();
-builder.Services.AddEndpointsApiExplorer();
-builder.Services.AddExceptionHandler<GlobalExceptionHandler>();
-builder.Services.AddProblemDetails();
 
-// Ensure DbContext and services are registered as scoped
-builder.Services.AddDbContext<TodoDbContext>(options =>
-    options.UseNpgsql(builder.Configuration.GetConnectionString("DefaultConnection"))
-);
-builder.Services.AddScoped<ITaskService, TaskService>();
-
-builder.Services.AddCors(options =>
-{
-    var allowedOrigins = builder.Configuration.GetSection("CorsOrigins").Get<string[]>() ?? [];
-    
-    options.AddPolicy("AllowedOrigins",
-        policy =>
-        {
-            policy.WithOrigins(allowedOrigins)
-                .AllowAnyHeader()
-                .AllowAnyMethod();
-        });
-});
+builder.AddApplicationServices();
 
 var app = builder.Build();
 
@@ -55,7 +34,6 @@ using (var scope = app.Services.CreateScope())
     var dbContext = scope.ServiceProvider.GetRequiredService<TodoDbContext>();
     dbContext.Database.Migrate();
 }
-
 
 app.MapGroup("/api")
 .MapTaskEndpoints()
